+++ hpc/nvidia/cuda/samples/cooperative.cu
@@ -1,19 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
 #include <stdio.h>
-#if defined(HAVE_HPCC)
-#include <hc_runtime.h>
-#include <hpcc_cooperative_groups.h>
-#include "cuda_adapter.h"
-#elif defined(HAVE_HIP)
-#include <hip/hip_runtime.h>
-#include <hip/hip_cooperative_groups.h>
-#include "cuda_adapter.h"
-#else
-#include <cuda_runtime.h>
-#include <cooperative_groups.h>
-#endif
+#include "cuda_compat.h"
 #include "cuda_helpers.h"
+
 
 namespace cg = cooperative_groups;