+++ hpc/nvidia/cuda/samples/cudaEvent.cu
@@ -6,15 +6,7 @@
 #include <stdbool.h>
 #include <unistd.h>
 #include <time.h>
-#if defined(HAVE_HPCC)
-# include <hc_runtime.h>
-# include "cuda_adapter.h"
-#elif defined(HAVE_HIP)
-#include <hip/hip_runtime.h>
-#include "cuda_adapter.h"
-#else
-# include <cuda_runtime.h>
-#endif
+#include "cuda_compat.h"
 #include "cuda_helpers.h"