+++ hpc/nvidia/cuda/samples/cudaGetDeviceCount.cu
@@ -1,17 +1,5 @@
 #include <stdio.h>
-#if defined(HAVE_HPCC)
-#include <hc_runtime.h>
-#include "cuda_adapter.h"
-#elif defined(HAVE_HIP)
-#include <hip/hip_runtime.h>
-#include "cuda_adapter.h"
-#else
-/**
- * Actually, nvcc will add -include cuda_runtime.h during compile, thus, we
- * don't need this include line
- */
-#include <cuda_runtime.h>
-#endif
+#include "cuda_compat.h"
 #include "cuda_helpers.h"