+++ hpc/nvidia/cuda/samples/cudaMallocManaged.cu
@@ -3,15 +3,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <sys/types.h>
-#if defined(HAVE_HPCC)
-#include <hc_runtime.h>
-#include "cuda_adapter.h"
-#elif defined(HAVE_HIP)
-#include <hip/hip_runtime.h>
-#include "cuda_adapter.h"
-#else
-#include <cuda_runtime.h>
-#endif
+#include "cuda_compat.h"
 #include "cuda_helpers.h"