+++ hpc/nvidia/cuda/samples/hello.cu
@@ -4,19 +4,7 @@
  * elf dump PTX and SASS: commit 306555e1b4e6 ("nvidia: PTX and SASS")
  */
 #include <stdio.h>
-#if defined(HAVE_HPCC)
-#include <hc_runtime.h>
-#include "cuda_adapter.h"
-#elif defined(HAVE_HIP)
-#include <hip/hip_runtime.h>
-#include "cuda_adapter.h"
-#else
-/**
- * Actually, nvcc will add -include cuda_runtime.h during compile, thus, we
- * don't need this include line
- */
-#include <cuda_runtime.h>
-#endif
+#include "cuda_compat.h"
 #include "cuda_helpers.h"